+++ dcpp/QueueManager.h
@@ -266,7 +266,7 @@ private:
     /** The queue needs to be saved */
     bool dirty;
     /** Next search */
-    uint32_t nextSearch;
+    uint64_t nextSearch;
     /** File lists not to delete */
     StringList protectedFileLists;
     /** Sanity check for the target filename */
@@ -290,8 +290,8 @@ private:
     uint32_t calcCrc32(const string& file) throw(FileException);
 
     // TimerManagerListener
-    virtual void on(TimerManagerListener::Second, uint32_t aTick) throw();
-    virtual void on(TimerManagerListener::Minute, uint32_t aTick) throw();
+    virtual void on(TimerManagerListener::Second, uint64_t aTick) throw();
+    virtual void on(TimerManagerListener::Minute, uint64_t aTick) throw();
 
     // SearchManagerListener
     virtual void on(SearchManagerListener::SR, const SearchResultPtr&) throw();
