+++ eiskaltdcpp-gtk/src/mainwindow.cc
@@ -2198,7 +2198,7 @@ void MainWindow::on(QueueManagerListener::Finished, QueueItem *item, const strin
     }
 }
 
-void MainWindow::on(TimerManagerListener::Second, uint32_t ticks) throw()
+void MainWindow::on(TimerManagerListener::Second, uint64_t ticks) throw()
 {
     // Avoid calculating status update if it's not needed
     if (!WGETB("always-tray") && minimized)
